Moved alerts to separate context
