+++ src/context/alert.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const AlertContext = React.createContext(() => {});
+
+export default AlertContext;
